refactor: remove grid
