+++ tests/test_mda_widget.py
@@ -126,7 +126,7 @@ def test_mda_grid(qtbot: QtBot, global_mmcore: CMMCorePlus):
     # w/o overlap
     grid_wdg.scan_size_spinBox_r.setValue(2)
     grid_wdg.scan_size_spinBox_c.setValue(2)
-    grid_wdg.ovelap_spinBox.setValue(0)
+    grid_wdg.overlap_spinBox.setValue(0)
     assert grid_wdg.info_lbl.text() == "1.024 mm x 1.024 mm"
 
     mock = Mock()
@@ -153,7 +153,7 @@ def test_mda_grid(qtbot: QtBot, global_mmcore: CMMCorePlus):
     # with overlap
     grid_wdg.scan_size_spinBox_r.setValue(3)
     grid_wdg.scan_size_spinBox_c.setValue(3)
-    grid_wdg.ovelap_spinBox.setValue(15)
+    grid_wdg.overlap_spinBox.setValue(15)
     assert grid_wdg.info_lbl.text() == "1.306 mm x 1.306 mm"
 
     grid_wdg.clear_checkbox.setChecked(False)
